"#00095 - improve scrape world file structure (move service, fix affected lines)"
